+++ src/byte/domain/agent/implementations/ask/agent.py
@@ -7,7 +7,7 @@ from byte.domain.agent.nodes.assistant_node import AssistantNode
 from byte.domain.agent.nodes.end_node import EndNode
 from byte.domain.agent.nodes.start_node import StartNode
 from byte.domain.agent.nodes.tool_node import ToolNode
-from byte.domain.agent.schemas import AssistantRunnable
+from byte.domain.agent.schemas import AssistantContextSchema
 from byte.domain.agent.state import AskState
 from byte.domain.llm.service.llm_service import LLMService
 from byte.domain.mcp.service.mcp_service import MCPService
@@ -40,30 +40,15 @@ class AskAgent(Agent):
 		Usage: `graph = await agent.build()`
 		"""
 
-		assistant_runnable = await self.get_assistant_runnable()
-
 		# Create the state graph
 		graph = StateGraph(self.get_state_class())
 
 		# Add nodes
-		graph.add_node(
-			"start_node",
-			await self.make(StartNode, agent=self.__class__.__name__),
-		)
-		graph.add_node(
-			"assistant_node",
-			await self.make(AssistantNode, runnable=assistant_runnable.runnable),
-		)
-		graph.add_node("tools_node", await self.make(ToolNode, tools=assistant_runnable.tools))  # pyright: ignore[reportArgumentType]
+		graph.add_node("start_node", await self.make(StartNode))
+		graph.add_node("assistant_node", await self.make(AssistantNode))
+		graph.add_node("tools_node", await self.make(ToolNode))
 
-		graph.add_node(
-			"end_node",
-			await self.make(
-				EndNode,
-				agent=self.__class__.__name__,
-				llm=assistant_runnable.llm,
-			),
-		)
+		graph.add_node("end_node", await self.make(EndNode))
 
 		# Define edges
 		graph.add_edge(START, "start_node")
@@ -77,16 +62,22 @@ class AskAgent(Agent):
 		checkpointer = await self.get_checkpointer()
 		return graph.compile(checkpointer=checkpointer)
 
-	async def get_assistant_runnable(self) -> AssistantRunnable:
+	async def get_assistant_runnable(self) -> AssistantContextSchema:
 		llm_service = await self.make(LLMService)
-		llm: BaseChatModel = llm_service.get_main_model()
+		main: BaseChatModel = llm_service.get_main_model()
+		weak: BaseChatModel = llm_service.get_weak_model()
 
 		mcp_service = await self.make(MCPService)
 		mcp_tools = mcp_service.get_tools_for_agent("ask")
 
-		# Create the assistant runnable with out any tools. So regardless it wont make a tool call even thou we have a tool node.
-		return AssistantRunnable(
-			runnable=ask_prompt | llm.bind_tools(mcp_tools, parallel_tool_calls=False),
-			llm=llm,
+		# test: RunnableSerializable[dict[Any, Any], BaseMessage] = ask_prompt | main
+		# main.bind_tools(mcp_tools, parallel_tool_calls=False)
+
+		return AssistantContextSchema(
+			mode="main",
+			prompt=ask_prompt,
+			main=main,
+			weak=weak,
+			agent=self.__class__.__name__,
 			tools=mcp_tools,
 		)