+++ src/byte/domain/agent/implementations/fixer/agent.py
@@ -2,7 +2,7 @@ from langchain_core.language_models.chat_models import BaseChatModel
 
 from byte.domain.agent.implementations.coder.agent import CoderAgent
 from byte.domain.agent.implementations.fixer.prompts import fixer_prompt
-from byte.domain.agent.schemas import AssistantRunnable
+from byte.domain.agent.schemas import AssistantContextSchema
 from byte.domain.llm.service.llm_service import LLMService
 
 
@@ -18,12 +18,12 @@ class FixerAgent(CoderAgent):
 	async def get_checkpointer(self):
 		return False
 
-	async def get_assistant_runnable(self) -> AssistantRunnable:
+	async def get_assistant_runnable(self) -> AssistantContextSchema:
 		llm_service = await self.make(LLMService)
 		llm: BaseChatModel = llm_service.get_main_model()
 
 		# Create the assistant runnable with out any tools. So regardless it wont make a tool call even thou we have a tool node.
-		return AssistantRunnable(
+		return AssistantContextSchema(
 			runnable=fixer_prompt | llm,
 			llm=llm,
 		)