+++ src/byte/domain/agent/nodes/end_node.py
@@ -1,28 +1,22 @@
 from langgraph.graph.state import END, RunnableConfig
+from langgraph.runtime import Runtime
 from langgraph.types import Command
 
 from byte.core.event_bus import EventType, Payload
 from byte.domain.agent.nodes.base_node import Node
+from byte.domain.agent.schemas import AssistantContextSchema
 from byte.domain.agent.state import BaseState
 
 
 class EndNode(Node):
-	async def boot(
-		self,
-		agent: str,
-		llm,
-		**kwargs,
-	):
-		self.agent = agent
-		self.llm = llm
-
-	async def __call__(self, state: BaseState, config: RunnableConfig):
+	async def __call__(self, state: BaseState, config: RunnableConfig, runtime: Runtime[AssistantContextSchema]):
 		payload = Payload(
 			event_type=EventType.END_NODE,
 			data={
 				"state": state,
-				"agent": self.agent,
-				"llm": self.llm,
+				"agent": runtime.context.agent,
+				# TODO: This needs to be updated to get the main llm for this node.
+				"llm": runtime.context.main,
 			},
 		)
 		await self.emit(payload)